HTTPBody writer and readers
