+++ larking/mux.go
@@ -8,7 +8,6 @@ import (
 	"bytes"
 	"context"
 	"crypto/sha256"
-	"encoding/json"
 	"fmt"
 	"io"
 	"math"
@@ -20,11 +19,9 @@ import (
 	"sync/atomic"
 	"time"
 
-	"github.com/gobwas/ws"
 	"google.golang.org/genproto/googleapis/api/annotations"
 	"google.golang.org/genproto/googleapis/api/httpbody"
 	"google.golang.org/genproto/googleapis/api/serviceconfig"
-	"google.golang.org/genproto/googleapis/rpc/code"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/metadata"
@@ -796,231 +793,6 @@ func (s *state) match(route, verb string) (*method, params, error) {
 		return nil, nil, status.Error(codes.NotFound, "not found")
 	}
 	return s.path.match(route, verb)
-}
-
-func isWebsocketRequest(r *http.Request) bool {
-	for _, header := range r.Header["Upgrade"] {
-		if header == "websocket" {
-			return true
-		}
-	}
-	return false
-}
-
-type twirpError struct {
-	Code    string            `json:"code"`
-	Message string            `json:"msg"`
-	Meta    map[string]string `json:"meta"`
-}
-
-func (m *Mux) encError(w http.ResponseWriter, r *http.Request, err error) {
-	s, _ := status.FromError(err)
-	if isTwirp := r.Header.Get("Twirp-Version") != ""; isTwirp {
-		accept := "application/json"
-
-		w.Header().Set("Content-Type", accept)
-		w.WriteHeader(HTTPStatusCode(s.Code()))
-
-		codeStr := strings.ToLower(code.Code_name[int32(s.Code())])
-
-		terr := &twirpError{
-			Code:    codeStr,
-			Message: s.Message(),
-		}
-		b, err := json.Marshal(terr)
-		if err != nil {
-			panic(err) // ...
-		}
-		w.Write(b) //nolint
-		return
-
-	}
-
-	accept := negotiateContentType(r.Header, m.opts.contentTypeOffers, "application/json")
-	c := m.opts.codecs[accept]
-
-	w.Header().Set("Content-Type", accept)
-	w.WriteHeader(HTTPStatusCode(s.Code()))
-
-	b, err := c.Marshal(s.Proto())
-	if err != nil {
-		panic(err) // ...
-	}
-	w.Write(b) //nolint
-}
-
-func (m *Mux) serveHTTP(w http.ResponseWriter, r *http.Request) error {
-	ctx, mdata := newIncomingContext(r.Context(), r.Header)
-
-	s := m.loadState()
-	isWebsocket := isWebsocketRequest(r)
-
-	verb := r.Method
-	if isWebsocket {
-		verb = kindWebsocket
-	}
-
-	method, params, err := s.match(r.URL.Path, verb)
-	if err != nil {
-		return err
-	}
-
-	queryParams, err := method.parseQueryParams(r.URL.Query())
-	if err != nil {
-		return err
-	}
-	params = append(params, queryParams...)
-
-	hd, err := s.pickMethodHandler(method.name)
-	if err != nil {
-		return err
-	}
-
-	// Handle stats.
-	beginTime := time.Now()
-	if sh := m.opts.statsHandler; sh != nil {
-		ctx = sh.TagRPC(ctx, &stats.RPCTagInfo{
-			FullMethodName: hd.method,
-			FailFast:       false, // TODO
-		})
-
-		sh.HandleRPC(ctx, &stats.InHeader{
-			FullMethod:  method.name,
-			RemoteAddr:  strAddr(r.RemoteAddr),
-			Compression: r.Header.Get("Content-Encoding"),
-			Header:      metadata.MD(mdata).Copy(),
-		})
-
-		sh.HandleRPC(ctx, &stats.Begin{
-			Client:                    false,
-			BeginTime:                 beginTime,
-			FailFast:                  false, // TODO
-			IsClientStream:            hd.desc.IsStreamingClient(),
-			IsServerStream:            hd.desc.IsStreamingServer(),
-			IsTransparentRetryAttempt: false, // TODO
-		})
-	}
-
-	if isWebsocket {
-		conn, _, _, err := ws.UpgradeHTTP(r, w)
-		if err != nil {
-			return err
-		}
-		defer conn.Close()
-
-		stream := &streamWS{
-			ctx:    ctx,
-			conn:   conn,
-			method: method,
-			params: params,
-		}
-		herr := hd.handler(&m.opts, stream)
-
-		if herr != nil {
-			s, _ := status.FromError(herr)
-			// TODO: limit message size.
-
-			code := WSStatusCode(s.Code())
-			f := ws.NewCloseFrame(ws.NewCloseFrameBody(code, s.Message()))
-			b, err := ws.CompileFrame(f)
-			if err != nil {
-				return err
-			}
-			if _, err := conn.Write(b); err != nil {
-				return err
-			}
-		} else {
-			if _, err := conn.Write(ws.CompiledClose); err != nil {
-				return err
-			}
-		}
-
-		// Handle stats.
-		if sh := m.opts.statsHandler; sh != nil {
-			endTime := time.Now()
-			sh.HandleRPC(ctx, &stats.End{
-				Client:    false,
-				BeginTime: beginTime,
-				EndTime:   endTime,
-				Error:     err,
-			})
-		}
-		return nil
-	}
-
-	contentType := r.Header.Get("Content-Type")
-	if contentType == "" {
-		contentType = "application/json"
-	}
-	contentEncoding := r.Header.Get("Content-Encoding")
-
-	var body io.Reader = r.Body
-	if cz := m.opts.compressors[contentEncoding]; cz != nil {
-		z, err := cz.Decompress(r.Body)
-		if err != nil {
-			return err
-		}
-		body = z
-	}
-
-	accept := negotiateContentType(r.Header, m.opts.contentTypeOffers, contentType)
-	acceptEncoding := negotiateContentEncoding(r.Header, m.opts.encodingTypeOffers)
-
-	var resp io.Writer = w
-	if cz := m.opts.compressors[acceptEncoding]; cz != nil {
-		w.Header().Set("Content-Encoding", acceptEncoding)
-		z, err := cz.Compress(w)
-		if err != nil {
-			return err
-		}
-		defer z.Close()
-		resp = z
-	}
-
-	stream := &streamHTTP{
-		ctx:    ctx,
-		method: method,
-		params: params,
-		opts:   m.opts,
-
-		// write
-		w:       resp,
-		wHeader: w.Header(),
-
-		// read
-		r:       body,
-		rHeader: r.Header,
-
-		contentType:    contentType,
-		accept:         accept,
-		acceptEncoding: acceptEncoding,
-		hasBody:        r.ContentLength > 0 || r.ContentLength == -1,
-	}
-	herr := hd.handler(&m.opts, stream)
-	// Handle stats.
-	if sh := m.opts.statsHandler; sh != nil {
-		endTime := time.Now()
-
-		// Try to send Trailers, might not be respected.
-		setOutgoingHeader(w.Header(), stream.trailer)
-		sh.HandleRPC(ctx, &stats.OutTrailer{
-			Trailer: stream.trailer.Copy(),
-		})
-
-		sh.HandleRPC(ctx, &stats.End{
-			Client:    false,
-			BeginTime: beginTime,
-			EndTime:   endTime,
-			Error:     herr,
-		})
-	}
-	if herr != nil {
-		if !stream.sentHeader {
-			w.Header().Set("Content-Encoding", "identity") // try to avoid gzip
-		}
-		m.encError(w, r, herr)
-	}
-	return nil
 }
 
 // ServeHTTP implements http.Handler.
